pyproject.toml
[project]
name = "dingding"
version = "0.1.0"
description = "Multi-task notification bot for DingTalk with weather, hot search and more"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.2",
    "croniter>=2.0.0",
    "apscheduler>=3.10.0"
]
